secure-fx-payment-gateway/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
redis==5.0.1
aiohttp==3.9.1
requests==2.31.0
asgiref==3.7.2
